database/webbservice/requirements/base.txt
Django==1.7.4
djangorestframework==3.0.4
Fabric==1.10.1
ecdsa==0.13
paramiko==1.15.2
pycrypto==2.6.1
wsgiref==0.1.2
Unipath==1.0